+++ src/components/layout/header/Header.tsx
@@ -5,12 +5,17 @@ import styles from "./Header.module.scss";
 
 interface Props {
     center?: boolean;
+    variant?: "toolbar";
+    position?: "top" | "left" | "right" | "bottom";
 }
 
-const Header: React.FC<Props> = ({ center, children }) => {
+const Header: React.FC<Props> = (props) => {
+    const { center, children, variant, position = "top" } = props;
+
     const cls = classnames(styles.header, {
         [styles["header--center"]]: center,
-    });
+        [styles["header--toolbar"]]: variant === "toolbar",
+    }, styles[`header--${position}`]);
 
     return (
         <div className={cls}>